+++ services/TestService/src/main/java/ro/uaic/fii/TestService/service/QuestionService.java
@@ -1,45 +1,69 @@
 package ro.uaic.fii.TestService.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ro.uaic.fii.TestService.dto.mapper.QuestionMapper;
+import ro.uaic.fii.TestService.dto.reqDto.QuestionReqDto;
+import ro.uaic.fii.TestService.dto.resDto.QuestionResDto;
 import ro.uaic.fii.TestService.exceptions.NotFoundException;
-import ro.uaic.fii.TestService.model.Question;
+import ro.uaic.fii.TestService.repository.model.Question;
 import ro.uaic.fii.TestService.repository.QuestionRepository;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class QuestionService {
+
     private final QuestionRepository questionRepository;
-
-    public QuestionService(QuestionRepository questionRepository) {
-        this.questionRepository = questionRepository;
+    private final QuestionMapper questionMapper;
+    public List<QuestionResDto> getAll() {
+        List<Question> questions = questionRepository.findAll();
+        return questions.stream().map(questionMapper::toDto).toList();
     }
 
-    public List<Question> getAll() {
-        return questionRepository.findAll();
+    public QuestionResDto getById(int id) {
+        Question question =  questionRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Question", id));
+        return questionMapper.toDto(question);
     }
 
-    public Question getById(Integer id) {
-        return  questionRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Question with id : " + id + " not found."));
+    public QuestionResDto save(QuestionReqDto dto) {
+        Question savedQuestion = questionRepository.save(questionMapper.dtoToEntity(dto));
+        return questionMapper.toDto(savedQuestion);
     }
 
-    public Question save(Question question) {
-        return questionRepository.save(question);
+    public QuestionResDto update(Integer id, QuestionReqDto dto) {
+        Question questionToUpdate = questionRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Question", id));
+        questionToUpdate.setDomainId(dto.getDomainId());
+        questionToUpdate.setContent(dto.getContent());
+        questionToUpdate.setActive(dto.isActive());
+        questionToUpdate.setNotes(dto.getNotes());
+        questionToUpdate.setUpdateUid(dto.getUserUid());
+
+        Question updatedQuestion = questionRepository.save(questionToUpdate);
+        return questionMapper.toDto(updatedQuestion);
     }
 
-    public Question update(Integer id, Question question) {
-        Question questionToUpdate = getById(id);
-        questionToUpdate.setDomainId(question.getDomainId());
-        questionToUpdate.setContent(question.getContent());
-        questionToUpdate.setActive(question.getActive());
-        questionToUpdate.setNotes(question.getNotes());
-        questionToUpdate.setUpdateUid(question.getUpdateUid());
-
-        return questionRepository.save(questionToUpdate);
-    }
-
-    public void delete(Integer id) {
+    public void delete(int id) {
+        if (!questionRepository.existsById(id)) {
+            throw new NotFoundException("Question", id);
+        }
         questionRepository.deleteById(id);
+    }
+
+    public void addTopicToQuestion(int id, int topicId) {
+        Question question = questionRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Question", id));
+        question.getTopicIds().add(topicId);
+        questionRepository.save(question);
+    }
+
+    public void removeTopicFromQuestion(int id, int topicId) {
+        Question question = questionRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Question", id));
+        question.getTopicIds().remove(topicId);
+        questionRepository.save(question);
     }
 }
