+++ services/TestService/src/main/java/ro/uaic/fii/TestService/service/ResponseService.java
@@ -1,48 +1,71 @@
 package ro.uaic.fii.TestService.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ro.uaic.fii.TestService.dto.mapper.ResponseMapper;
+import ro.uaic.fii.TestService.dto.reqDto.ResponseReqDto;
+import ro.uaic.fii.TestService.dto.resDto.ResponseResDto;
+import ro.uaic.fii.TestService.exceptions.BadRequestException;
 import ro.uaic.fii.TestService.exceptions.NotFoundException;
-import ro.uaic.fii.TestService.model.Response;
+import ro.uaic.fii.TestService.repository.model.Response;
+import ro.uaic.fii.TestService.repository.ParticipantRepository;
+import ro.uaic.fii.TestService.repository.QuestionRepository;
 import ro.uaic.fii.TestService.repository.ResponseRepository;
 
 import java.util.List;
 
+import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
+
 @Service
+@RequiredArgsConstructor
 public class ResponseService {
 
     private final ResponseRepository responseRepository;
+    private final ParticipantRepository participantRepository;
+    private final QuestionRepository questionRepository;
+    private final ResponseMapper responseMapper;
 
-    public ResponseService(ResponseRepository responseRepository) {
-        this.responseRepository = responseRepository;
+    public List<ResponseResDto> getAll() {
+        List<Response> responses = responseRepository.findAll();
+        return responses.stream().map(responseMapper::toDto).toList();
     }
 
-    public List<Response> getAll() {
-        return responseRepository.findAll();
+    public ResponseResDto getById(int id) {
+        Response response =  responseRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Response", id));
+        return responseMapper.toDto(response);
     }
 
-    public Response getById(Integer id) {
-        return  responseRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Response with id : " + id + " not found."));
+    public ResponseResDto save(ResponseReqDto dto) {
+        Response savedResponse = responseRepository.save(responseMapper.dtoToEntity(dto));
+        return responseMapper.toDto(savedResponse);
     }
 
-    public Response save(Response response) {
-        return responseRepository.save(response);
-    }
-
-    public Response update(Integer id, Response response) {
-        Response responseToUpdate = getById(id);
-        responseToUpdate.setParticipantId(response.getParticipantId());
-        responseToUpdate.setOrderNumber(response.getOrderNumber());
-        responseToUpdate.setBaseQuestionId(response.getBaseQuestionId());
-        responseToUpdate.setQuestion(response.getQuestion());
-        responseToUpdate.setAnswer(response.getAnswer());
-        responseToUpdate.setPoints(response.getPoints());
-        responseToUpdate.setIpAddress(response.getIpAddress());
-        responseToUpdate.setUpdateUid(response.getUpdateUid());
-        return save(responseToUpdate);
+    public ResponseResDto update(int id, ResponseReqDto dto) {
+        Response responseToUpdate = responseRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Response", id));
+        if (!participantRepository.existsById(dto.getParticipantId())) {
+            throw new BadRequestException("Participant", dto.getParticipantId());
+        }
+        if (isNotEmpty(dto.getBaseQuestionId()) && !questionRepository.existsById(dto.getBaseQuestionId())) {
+            throw new BadRequestException("Question", dto.getBaseQuestionId());
+        }
+        responseToUpdate.setParticipantId(dto.getParticipantId());
+        responseToUpdate.setOrderNumber(dto.getOrderNumber());
+        responseToUpdate.setBaseQuestionId(dto.getBaseQuestionId());
+        responseToUpdate.setQuestion(dto.getQuestion());
+        responseToUpdate.setAnswer(dto.getAnswer());
+        responseToUpdate.setPoints(dto.getPoints());
+        responseToUpdate.setIpAddress(dto.getIpAddress());
+        responseToUpdate.setUpdateUid(dto.getUserUid());
+        Response updatedResponse = responseRepository.save(responseToUpdate);
+        return responseMapper.toDto(updatedResponse);
     }
 
     public void delete(Integer id) {
+        if (!responseRepository.existsById(id)) {
+            throw new NotFoundException("Response", id);
+        }
         responseRepository.deleteById(id);
     }
 }
